+++ examples/crud-api/CHANGELOG.md
@@ -1,5 +1,12 @@
 # celsian-crud-api-example
 
+## 1.0.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @celsian/core@0.6.3
+
 ## 1.0.11
 
 ### Patch Changes