+++ packages/celsian/CHANGELOG.md
@@ -1,5 +1,13 @@
 # celsian
 
+## 0.6.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @celsian/core@0.6.3
+  - @celsian/schema@0.6.3
+
 ## 0.6.2
 
 ### Patch Changes